+++ xenon-common/src/main/java/com/vmware/xenon/common/ServiceHost.java
@@ -66,6 +66,7 @@ import java.util.logging.Logger;
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManagerFactory;
 
+import io.netty.handler.codec.http.cors.CorsConfig;
 import io.opentracing.ActiveSpan;
 import io.opentracing.ActiveSpan.Continuation;
 import io.opentracing.SpanContext;
@@ -1556,6 +1557,7 @@ public class ServiceHost implements ServiceRequestSender {
                 this.httpListener.setResponsePayloadSizeLimit(this.state.responsePayloadSizeLimit);
             }
 
+            this.httpListener.setCorsConfig(getCorsConfig());
             this.httpListener.start(getPort(), this.state.bindAddress);
         }
 
@@ -1579,6 +1581,7 @@ public class ServiceHost implements ServiceRequestSender {
                             .setResponsePayloadSizeLimit(this.state.responsePayloadSizeLimit);
                 }
 
+                this.httpsListener.setCorsConfig(getCorsConfig());
                 this.httpsListener.start(getSecurePort(), this.state.bindAddress);
             }
         }
@@ -5783,5 +5786,29 @@ public class ServiceHost implements ServiceRequestSender {
      */
     public Tracer getTracer() {
         return this.otTracer;
+    }
+
+    /**
+     * CORS configuration for netty pipeline.
+     *
+     * By default, CORS is disabled.
+     * To enable CORS support, override this method in subclass and populate {@link CorsConfig}.
+     *
+     * Example:
+     * {@code
+     *   // allow PUT for any origin
+     *   CorsConfigBuilder.forAnyOrigin().allowedRequestMethods(HttpMethod.PUT).build();
+     * }
+     *
+     * When {@code null} is returned, CORS support from nettty(by {@link io.netty.handler.codec.http.cors.CorsHandler})
+     * will be disabled.
+     *
+     * @return CORS configuration
+     * @see io.netty.handler.codec.http.cors.CorsConfigBuilder
+     * @see io.netty.handler.codec.http.cors.CorsConfig
+     */
+    public CorsConfig getCorsConfig() {
+        // CORS is disabled by default
+        return null;
     }
 }