+++ xenon-common/src/main/java/com/vmware/xenon/common/ServiceRequestListener.java
@@ -16,6 +16,8 @@ package com.vmware.xenon.common;
 import java.io.IOException;
 import java.net.URI;
 
+import io.netty.handler.codec.http.cors.CorsConfig;
+
 import com.vmware.xenon.common.config.XenonConfiguration;
 
 public interface ServiceRequestListener {
@@ -54,4 +56,6 @@ public interface ServiceRequestListener {
     void setSecureAuthCookie(boolean secureAuthCookie);
 
     boolean getSecureAuthCookie();
+
+    void setCorsConfig(CorsConfig corsConfig);
 }